qemu_arm_mps2: uart is safe now that panic writer is safe
With `create_panic_writer` no longer an `unsafe fn`, `UartPanicWriter`
contains no `unsafe` at all, so it belongs in `qemu_arm_mps2` next to the
`Uart` it wraps rather than in `qemu_arm_mps2-unsafe`. `reset()` goes back
to private now that its only caller is in the same module; the boards
still take `UART0_BASE` from the unsafe crate.
