+++ boards/qemu_arm_mps2/an386/Cargo.toml
@@ -12,6 +12,7 @@ edition.workspace = true
 [dependencies]
 cortexm4 = { path = "../../../arch/cortex-m4" }
 kernel = { path = "../../../kernel" }
+qemu_arm_mps2 = { path = "../../../chips/qemu_arm_mps2" }
 qemu_arm_mps2-unsafe = { path = "../../../chips/qemu_arm_mps2-unsafe" }
 qemu_arm_mps2_an386 = { path = "../../../chips/qemu_arm_mps2_an386" }
 mps2_base = { path = "../mps2_base" }
