+++ chips/qemu_arm_mps2-unsafe/README.md
@@ -2,4 +2,4 @@ ARM MPS2 "Chip" Family - Unsafe
 ===============================
 
 This contains the `unsafe` code for the `qemu_arm_mps2` chip: the machine's
-memory map, the `Chip` implementation, and the panic writer.
+memory map and the `Chip` implementation.
